requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
streamlit
chromadb
pypdf
python-dotenv
tiktoken
requests
pandas
openpyxl
